# backend/internal/handler/v1/swagger/swagger.yaml
definitions:
...
  dd-prediction-api_internal_model.DuelTransaction:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      duelID:
        type: string
      projectID:
        type: string
      signature:
        type: string
      txType:
        type: integer
      userID:
        type: string
    type: object
...
  dd-prediction-api_internal_model.PartnerDailyIncome:
    properties:
      amount_usd:
        type: number
      day:
        type: string
    type: object
  dd-prediction-api_internal_model.PartnerDashboard:
    properties:
      api_key:
        type: string
      duel_count:
        type: integer
      monthly_volume_usd:
        type: number
      net_by_symbol:
        items:
          $ref: '#/definitions/dd-prediction-api_internal_model.SymbolNet'
        type: array
      platform_rate:
        type: number
      status:
        $ref: '#/definitions/dd-prediction-api_internal_model.ProjectStatus'
      total_volume_usd:
        type: number
    type: object
  dd-prediction-api_internal_model.PartnerMonthlyDDCommission:
    properties:
      dd_amount_usd:
        type: number
      dd_rate:
        type: number
      gross_usd:
        type: number
      month:
        type: string
    type: object
  dd-prediction-api_internal_model.PartnerMonthlyMAU:
    properties:
      active_users:
        type: integer
      month:
        type: string
    type: object
...
  dd-prediction-api_internal_model.ProjectCommissionClaim:
    properties:
      amount_usd:
        type: number
      claimed_at:
        type: string
      gross_usd:
        type: number
      id:
        type: string
      period_end:
        type: string
      period_start:
        type: string
      platform_rate:
        type: number
      project_id:
        type: string
      tx_hashes:
        additionalProperties:
          type: string
        type: object
      type:
        type: string
      wallet_address:
        type: string
    type: object
  dd-prediction-api_internal_model.ProjectCommissionClaimReq:
    properties:
      wallet_address:
        type: string
    type: object
  dd-prediction-api_internal_model.ProjectStatus:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - ProjectStatusActive
    - ProjectStatusDisabled
    - ProjectStatusBlocked
...
  dd-prediction-api_internal_model.SymbolNet:
    properties:
      gross_usd:
        type: number
      net_raw:
        type: integer
      net_usd:
        type: number
      platform_rate:
        type: number
      symbol:
        type: string
    type: object
...
  dd-prediction-api_pkg_apperrors.ErrorPublic:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a swagger specification for a Duel Duck back-end.
  title: Duel Duck Prediction API
  version: "1.0"
paths:
...
  /me/analytics/daily-income:
    get:
      description: 'Returns daily commission income (USDC equivalent) for the partner''s
        project within a date range. Query params: from, to (format: 2006-01-02).
        Defaults to current month.'
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Project API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Daily income list
          schema:
            items:
              $ref: '#/definitions/dd-prediction-api_internal_model.PartnerDailyIncome'
            type: array
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Partner permission required
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Get daily income analytics
      tags:
      - partner
  /me/analytics/dd-commission-monthly:
    get:
      description: Returns per-billing-month DD commission amounts and rates, derived
        from commission accruals.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Project API key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Monthly DD commission breakdown
          schema:
            items:
              $ref: '#/definitions/dd-prediction-api_internal_model.PartnerMonthlyDDCommission'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Partner permission required
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Get DD commission breakdown by month
      tags:
      - partner
  /me/analytics/monthly-active-users:
    get:
      description: Returns unique user counts per calendar month. A user is active
        if they placed at least one prediction in that month.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Project API key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Monthly active user counts
          schema:
            items:
              $ref: '#/definitions/dd-prediction-api_internal_model.PartnerMonthlyMAU'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Partner permission required
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Get monthly active users analytics
      tags:
      - partner
  /me/commission/claim:
    post:
      consumes:
      - application/json
      description: Claims all accumulated partner commission for completed billing
        periods and sends it to the provided wallet address.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Wallet address to send commission to
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dd-prediction-api_internal_model.ProjectCommissionClaimReq'
      produces:
      - application/json
      responses:
        "200":
          description: Created commission claim record
          schema:
            $ref: '#/definitions/dd-prediction-api_internal_model.ProjectCommissionClaim'
        "400":
          description: Invalid request or nothing to claim
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Partner permission required
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Claim partner commission
      tags:
      - partner
  /me/commission/history:
    get:
      description: Returns the full history of partner commission claims for the authenticated
        user's project.
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of commission claim records
          schema:
            items:
              $ref: '#/definitions/dd-prediction-api_internal_model.ProjectCommissionClaim'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Partner permission required
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Get partner commission claim history
      tags:
      - partner
  /me/dashboard:
    get:
      description: 'Returns the partner''s project dashboard: status, total/monthly
        volume, commission breakdown by symbol, and net profit after DD fee.'
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Project API key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Partner dashboard metrics
          schema:
            $ref: '#/definitions/dd-prediction-api_internal_model.PartnerDashboard'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Partner permission required
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Get partner dashboard
      tags:
      - partner
...
  /me/transactions:
    get:
      description: Returns on-chain transactions associated with the partner's project
        (predictions, rewards, refunds, commissions).
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Project API key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction list
          schema:
            items:
              $ref: '#/definitions/dd-prediction-api_internal_model.DuelTransaction'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "403":
          description: Partner permission required
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      security:
      - BearerAuth: []
      summary: Get project transactions
      tags:
      - partner
...
securityDefinitions:
  BearerAuth:
    description: Enter the auth token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"